+++ framework/model/src/main/java/is/codion/framework/model/package-info.java
@@ -20,7 +20,6 @@
  * Package configuration values:
  * <ul>
  * <li>{@link is.codion.framework.model.EntityApplicationModel#USERNAME_PREFIX}
- * <li>{@link is.codion.framework.model.EntityApplicationModel#AUTHENTICATION_REQUIRED}
  * <li>{@link is.codion.framework.model.EntityApplicationModel#SAVE_DEFAULT_USERNAME}
  * <li>{@link is.codion.framework.model.EntityEditModel#POST_EDIT_EVENTS}
  * <li>{@link is.codion.framework.model.EntityEditModel.EntityEditor#PERSIST_FOREIGN_KEYS}