+++ swing/framework-ui/src/test/java/is/codion/swing/framework/ui/EntityApplicationPanelTest.java
@@ -75,7 +75,6 @@ public class EntityApplicationPanelTest {
 						.domainType(TestDomain.DOMAIN)
 						.setUncaughtExceptionHandler(false)
 						.saveDefaultUsername(false)
-						.loginRequired(false)
 						.displayFrame(false)
 						.includeMainMenu(true)
 						.displayStartupDialog(false)