+++ changelog.md
@@ -38,6 +38,8 @@ Codion Change Log
 - EntityApplicationPanel.Builder.LoginProvider removed, loginProvider() replaced with userSupplier().
 - EntityApplicationPanel.Builder.ConnectionProviderFactory removed, Builder.connectionProvider() added, related refactoring.
 - EntityApplicationPanel.Builder.defaultLoginUser(), automaticLoginUser() and userSupplier() renamed defaultUser(), user() and user() respectively.
+- EntityApplicationPanel.Builder.loginRequired() removed.
+- EntityApplicationPanel.Builder.connectionProvider(EntityConnectionProvider) and user(User) added, connectionProvider(EntityConnectionProvider.Builder) replaced with connectionProvider(Function<User, EntityConnectionProvider>).
 ### is.codion.tools.generator.model
 - DomainGeneratorModel now supports databases without schemas and users, as in, SQLite.
 ### is.codion.tools.generator.ui